// src/api-client/models/budget-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Financial Aide Backend
 * API Backend for Financial Aide open-source budgeting system
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IntervalEnum } from './interval-enum';
import { UserResponse } from './user-response';
/**
 * 
 * @export
 * @interface BudgetResponse
 */
export interface BudgetResponse {
    /**
     * 
     * @type {string}
     * @memberof BudgetResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof BudgetResponse
     */
    description?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof BudgetResponse
     */
    start_time: Date;
    /**
     * 
     * @type {Date}
     * @memberof BudgetResponse
     */
    end_time: Date;
    /**
     * 
     * @type {IntervalEnum}
     * @memberof BudgetResponse
     */
    interval: IntervalEnum;
    /**
     * 
     * @type {string}
     * @memberof BudgetResponse
     */
    income: string;
    /**
     * 
     * @type {UserResponse}
     * @memberof BudgetResponse
     */
    user: UserResponse;
    /**
     * 
     * @type {Date}
     * @memberof BudgetResponse
     */
    created_at: Date;
    /**
     * 
     * @type {Date}
     * @memberof BudgetResponse
     */
    updated_at: Date;
    /**
     * 
     * @type {number}
     * @memberof BudgetResponse
     */
    id: number;
}